Web-scraper and data files
Java app using the web-scraper package Jaunt to download the data for the x-rays we want to work with and label it.  Also includes the products: the raw data in a huge json file, a labelled summary of each record including the major diagnosis in a smaller json file, and a small csv file with just the record number, image filename, and label for each record.